disable single-sign-out requests.
Our CASino usage does not rely on single-sign-out requests. Instead
logout is performed by removing the service cookies directly in the
`/logout` action. This works because the CASino service and every
service using it as an identity provider is hosted on the same domain.

Making these unnecessary requests slows down response time of our CASino
deployments and procudes unnecessary load for client services.