prevent Auto-Login handler from aborting redirects.
There is a possible race condition between Redirect after a successful
login and the JS Auto-Login timeout. This has the effect that the
Auto-Login handler aborts the redirect and initiates another request to
`/login` which will immediately redirect to the service.

The steps to reproduce are as follows:

1.) access service URL without a session
2.) GET /login
3.) login using valid credentials (POST /login)
4.) after logging in a `tgt` cookie will be set and the browser follows
the redirect back to service URL. If during that redirect the Auto-Login
handler fires, the redirect is aborted.
5.) GET /login (valid session exists)
6.) redirect to service URL.

This bug results in duplicated requests back to service URL (with
different `ticket` parameters). As a consequence `/serviceValidate` is
called multiple times as well.

This patch clears the Auto-Login timeout when the user submits the login
form.